+++ custom_components/wemportal/sensor.py
@@ -520,7 +520,6 @@ class WemPortalSensor(WemPortalEntity, RestoreSensor):
         except KeyError:
             self._attr_native_value = None
             _LOGGER.warning("Can't find %s", self._attr_unique_id)
-            _LOGGER.debug("Sensor data %s", self.coordinator.data)
 
         self.async_write_ha_state()